Code: Local Network ID and Port Scanner
Local network ID uses method returnIPNet() to return the network of the local machine : Add from LocalIP import returnIPNet to use

Port Scanner has method findOpenPorts(port, scanRange = 255,network = "no input") : Add from PortScan import Port Scanner

This method takes in a port number and OPTIONALLY a scan range and a network IP that ends in .0 such as 10.1.1.0 

It returns a list of machine IPs that have that particular port open.

scanRange: can me uses to restrict the number of addresses in the network that are scanned

network: if no value is entered for network the network will scan the local network. If a value is entered it will scan THAT network.